refactor: rename realtime server + move docker file
