+++ ci/pipeline.yml
@@ -54,6 +54,7 @@ jobs:
       - name: image
       params:
         CONTEXT: repo
+        DOCKERFILE: "repo/src/servers/realtime/Dockerfile"
       run:
         path: build
   - put: edge-image
